move oss operation functions to oss.py
