+++ python/runtime/pai/train_keras.py
@@ -18,8 +18,8 @@ from os import path
 
 import six
 import tensorflow as tf
+from runtime import oss
 from runtime.model_metadata import save_model_metadata
-from runtime.pai import model
 from runtime.pai.pai_distributed import (
     dump_into_tf_config, make_distributed_info_without_evaluator)
 from runtime.seeding import get_tf_random_seed
@@ -60,8 +60,8 @@ def keras_train_and_save(estimator, model_params, save, FLAGS,
                              has_none_optimizer)
 
     print("saving keras model to: %s" % FLAGS.sqlflow_oss_modeldir)
-    model.save_file(FLAGS.sqlflow_oss_modeldir, save)
-    model.save_file(FLAGS.sqlflow_oss_modeldir, "model_meta.json")
+    oss.save_file(FLAGS.sqlflow_oss_modeldir, save)
+    oss.save_file(FLAGS.sqlflow_oss_modeldir, "model_meta.json")
 
 
 def keras_train_distributed(classifier,
